Add validation for required fields in components (connectors a-c) (#800)
* New prompt - bump bundle json

* Initial plan

* Add validation for required fields in connectors a-c


* Fix trailing spaces from linter


* Change version bumps from minor to patch and update tests


---------

+++ src/appmixer/aws/s3/DeleteBucket/DeleteBucket.js
@@ -12,6 +12,11 @@ module.exports = {
         const { s3 } = commons.init(context);
 
         const { bucket } = context.messages.in.content;
+        if (!bucket) {
+            throw new context.CancelError('Bucket is required');
+        }
+
+
         await s3.deleteBucket({ Bucket: bucket }).promise();
 
         return context.sendJson({ Name: bucket }, 'deleted');
